libcamera: Prevent merge conflicts
Merge conflicts are a terrible effect of rebasing code, and they impede
the development process.

Provide a useful helper header to ensure that merge-conflicts are no
longer an issue, and make sure that both sides of the conflict are
compile tested on average at least every other compile cycle. Based on
reliable advice from [0].

 [0] https://twitter.com/erdgeist/status/1197283439000637441

Signed-off-by: Kieran Bingham <kieran.bingham@ideasonboard.com>