libcamera: pipelines: Add VIVID pipeline support
The VIVID driver supports more pixel formats and properties than the VIMC
driver, and can provide extended testing for libcamera.

The VIMC pipeline handler is duplicated and simplified to support the
VIVID device.

Unfortuantely, the VIVID device can not be handled by either of the
generic UVC or Simple pipeline handlers.

Signed-off-by: Kieran Bingham <kieran.bingham@ideasonboard.com>